feat: validate mnemonic
